Added timeoutDuration logic to trajectory following. Fixed cycle count issue. Added example trajectories for all common control signals

+++ include/adjustableSettings.h
@@ -15,7 +15,7 @@ extern const int ANALOG_VENT_MIN;
 extern const int ANALOG_VENT_MAX;
 
 // Frequency variables
-extern const int PRESSURE_READ_DELAY;
+extern const double PRESSURE_READ_DELAY;
 extern const int INTERP_CALC_DELAY;
 extern const int CONTROLLER_DELAY;
